Add files via upload
This Project is Based on Deep Learning Model. In this we are using BERT model. BERT is an open source machine learning framework for natural language processing (NLP). BERT is designed to help computers understand the meaning of ambiguous language in text by using surrounding text to establish context.
Main.py contains the code.
requirements.txt contains the required library to run the code.